# spec/api.spec.yaml
openapi: 3.0.0
paths:
  /{dataset}/{version}/records:
    post:
      summary: 'Provides search capability.'
      parameters:
        - name: version
          in: path
          description: Version of the dataset.
          required: true
          schema:
            type: string
            default: v1
        - name: dataset
          in: path
          description: 'Name of the dataset.'
          required: true
          schema:
            type: string
            default: oa_citations
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                criteria:
                  type: string
                  description: 'Search string'
                start:
                  description: Starting record number.
                  type: integer
                  minimum: 0
                rows:
                  type: integer
                  default: 100
                  minimum: 1
                  maximum: 100
              required:
                - criteria
                - start
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: 
                      type: number
                      example: 4011,
                    title:
                      type: string
                      example: 'Latest report 2019'
                    created:
                      type: string
                      format: date-time
                    archived:
                      type: boolean
                      default: false
                    ISBN:
                      type: string
                      pattern: 'ISBN\x20(?=.{13}$)\d{1,5}([- ])\d{1,7}\1\d{1,6}\1(\d|X)$'
                      example: 'ISBN 1-56389-016-X'
                    
        '404':
          description: No matching record found for the given criteria.